+++ core/src/test/java/io/questdb/test/griffin/ExplainPlanTest.java
@@ -1865,7 +1865,7 @@ public class ExplainPlanTest extends AbstractCairoTest {
 
                 for (int p = 0; p < sigArgCount; p++) {
                     int sigArgTypeMask = descriptor.getArgTypeMask(p);
-                    final byte sigArgType = FunctionFactoryDescriptor.toTypeTag(sigArgTypeMask);
+                    final short sigArgType = FunctionFactoryDescriptor.toTypeTag(sigArgTypeMask);
                     boolean isArray = FunctionFactoryDescriptor.isArray(sigArgTypeMask);
 
                     if (p > 0) {
@@ -1883,7 +1883,7 @@ public class ExplainPlanTest extends AbstractCairoTest {
                 for (int p = 0; p < sigArgCount; p++) {
                     int argTypeMask = descriptor.getArgTypeMask(p);
                     boolean isConstant = FunctionFactoryDescriptor.isConstant(argTypeMask);
-                    byte sigArgType = FunctionFactoryDescriptor.toTypeTag(argTypeMask);
+                    short sigArgType = FunctionFactoryDescriptor.toTypeTag(argTypeMask);
                     ObjList<Function> availableValues = constFuncs.get(sigArgType);
                     int constValues = availableValues != null ? availableValues.size() : 1;
                     combinations *= (constValues + (isConstant ? 0 : 1));
@@ -1900,7 +1900,7 @@ public class ExplainPlanTest extends AbstractCairoTest {
                     try {
                         for (int p = 0; p < sigArgCount; p++) {
                             int sigArgTypeMask = descriptor.getArgTypeMask(p);
-                            byte sigArgType = FunctionFactoryDescriptor.toTypeTag(sigArgTypeMask);
+                            short sigArgType = FunctionFactoryDescriptor.toTypeTag(sigArgTypeMask);
                             boolean isConstant = FunctionFactoryDescriptor.isConstant(sigArgTypeMask);
                             boolean isArray = FunctionFactoryDescriptor.isArray(sigArgTypeMask);
                             boolean useConst = isConstant || (tempNo & 1) == 1 || sigArgType == ColumnType.CURSOR || sigArgType == ColumnType.RECORD;
